(nextage_ros_bridge) Refactoring to separate modules per hand type, to allow more flexible hand tool combination. Not tested yet on a real robot and on simulation it isn't possible to test as of the moment.
(nextage_ros_bridge) more refactoring to optimize the classes that represent hands.
